adding static assets
Note: the dark/light logos are mostly superceded by the 'generic' ones controlled by styling, except for a couple of cases where I don't think I can apply styling... in the navbar, for instance, or the favicon. So I've left them all in there for now, in case I decide to use them in more of those places. Will try to remember to do a big cleanup eventually.